implement add methods script

+++ src/main/java/br/ufpa/labes/spm/service/dto/RoleNeedsAbilityDTO.java
@@ -1,7 +1,7 @@
 package br.ufpa.labes.spm.service.dto;
 
 public class RoleNeedsAbilityDTO {
-	private Integer oid;
+	private Long oid;
 
 	private Integer degree;
